fix(temperature): change datatype from float to double

+++ src/main/java/org/vincent/TemperatureUtils.java
@@ -1,7 +1,7 @@
 package org.vincent;
 
 public class TemperatureUtils {
-    public static float fahrenheitToCelsiusConverter(float temperature){
+    public static double fahrenheitToCelsiusConverter(double temperature){
         return ((temperature-32) * ((float) 5 /9));
     }
 }
